refactor(nostr): extract group-messaging cluster into useGroupMessaging (#707)
Move the NIP-17 group send (sendGroupMessage) and the kind-30200
group-state publish (publishGroupState) out of NostrProvider into a
dedicated useGroupMessaging hook, mirroring the useDmInbox extraction.

Behavior-preserving: the callbacks keep identical bodies, comments, and
[pubkey, isLoggedIn, signerType, relays] dep arrays; the provider threads
those deps in as arguments and re-exposes the returned callbacks. The
context value object is byte-for-byte unchanged.

NostrContext.tsx: 2031 -> 1879 lines; baseline lowered accordingly.
